# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AppCode:
    enum:
    - 0
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    type: integer
    x-enum-comments:
      CodeAuthRequired: Request requires authentication; no valid token provided.
      CodeBusinessLogic: Valid request but invalid in the current context.
      CodeConflict: Duplicate resource or incompatible state.
      CodeInternalError: Unexpected server-side error.
      CodeNotFound: Requested resource does not exist.
      CodePermissionDenied: Authenticated but lacks permission to perform this action.
      CodeRateLimited: Too many requests; client should back off.
      CodeSuccess: Request processed successfully.
      CodeValidationError: 'Invalid input: missing field, wrong type, or format violation.'
    x-enum-descriptions:
    - Request processed successfully.
    - 'Invalid input: missing field, wrong type, or format violation.'
    - Request requires authentication; no valid token provided.
    - Authenticated but lacks permission to perform this action.
    - Requested resource does not exist.
    - Duplicate resource or incompatible state.
    - Valid request but invalid in the current context.
    - Too many requests; client should back off.
    - Unexpected server-side error.
    x-enum-varnames:
    - CodeSuccess
    - CodeValidationError
    - CodeAuthRequired
    - CodePermissionDenied
    - CodeNotFound
    - CodeConflict
    - CodeBusinessLogic
    - CodeRateLimited
    - CodeInternalError
  dto.ArticleImageRequest:
    properties:
      image:
        type: string
      sort_order:
        type: integer
    required:
    - image
    type: object
  dto.ArticleImageResponse:
    properties:
      article_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      image:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
  dto.ArticleRequest:
    properties:
      content:
        type: string
      external_id:
        type: string
      label:
        type: string
      localization:
        type: object
      location_id:
        type: string
      navigate:
        type: string
      placement:
        type: string
      published_at:
        type: string
      published_period_end:
        type: string
      published_period_start:
        type: string
      status:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  dto.ArticleResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/dto.ArticleImageResponse'
        type: array
      label:
        type: string
      localization:
        type: object
      location_id:
        type: string
      navigate:
        type: string
      placement:
        type: string
      published_at:
        type: string
      published_period_end:
        type: string
      published_period_start:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      venue_id:
        type: string
    type: object
...
  dto.PaginatedData:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
  dto.PaginationMeta:
    properties:
      current_page:
        type: integer
      has_next_page:
        type: boolean
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  dto.Response:
    properties:
      code:
        $ref: '#/definitions/dto.AppCode'
      data: {}
      messages:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Digimap backend REST API
  title: Digimap Backend API
  version: "1.0"
paths:
...
  /venues/{id}/articles:
    get:
      description: List articles for a venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create a new article (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Article details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create article
      tags:
      - articles
  /venues/{id}/articles/{articleID}:
    delete:
      description: Delete an article (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Article ID
        in: path
        name: articleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete article
      tags:
      - articles
    get:
      description: Get an article by ID
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Article ID
        in: path
        name: articleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get article
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Update an article (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Article ID
        in: path
        name: articleID
        required: true
        type: string
      - description: Article details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update article
      tags:
      - articles
  /venues/{id}/articles/{articleID}/images:
    post:
      consumes:
      - application/json
      description: Add an image to an article (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Article ID
        in: path
        name: articleID
        required: true
        type: string
      - description: Image details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleImageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleImageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Add article image
      tags:
      - articles
  /venues/{id}/articles/{articleID}/images/{imageID}:
    delete:
      description: Delete an image from an article (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Article ID
        in: path
        name: articleID
        required: true
        type: string
      - description: Image ID
        in: path
        name: imageID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete article image
      tags:
      - articles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"